🧩Refactor: 架构规范性梳理和God Classes拆分
