+++ KitX Core Contracts/KitX.Core.Contract/Security/IEncryptionService.cs
@@ -1,67 +1,13 @@
-﻿using System;
-using System.Collections.Generic;
-using System.ComponentModel;
 using System.Threading.Tasks;
-using KitX.Shared.CSharp.Device;
 using KitX.Shared.CSharp.Security;
-using KitXIDeviceKey = KitX.Core.Contract.Configuration.IDeviceKey;
 
 namespace KitX.Core.Contract.Security;
 
 /// <summary>
-/// Security management service interface
+/// Encryption service interface
 /// </summary>
-public interface ISecurityService
+public interface IEncryptionService
 {
-    /// <summary>
-    /// Gets all device keys
-    /// </summary>
-    IReadOnlyList<KitXIDeviceKey> GetDeviceKeys();
-
-    /// <summary>
-    /// Adds a device key
-    /// </summary>
-    /// <param name="macAddress">The MAC address</param>
-    /// <param name="deviceName">The device name</param>
-    /// <param name="publicKey">The public key</param>
-    /// <returns>True if addition was successful</returns>
-    bool AddDeviceKey(string macAddress, string deviceName, string publicKey);
-
-    /// <summary>
-    /// Removes a device key
-    /// </summary>
-    /// <param name="macAddress">The MAC address</param>
-    /// <returns>True if removal was successful</returns>
-    bool RemoveDeviceKey(string macAddress);
-
-    /// <summary>
-    /// Searches for a device key by device locator
-    /// </summary>
-    /// <param name="locator">The device locator</param>
-    /// <returns>The device key if found, otherwise null</returns>
-    DeviceKey? SearchDeviceKey(DeviceLocator locator);
-
-    /// <summary>
-    /// Checks if a device key is correct
-    /// </summary>
-    /// <param name="locator">The device locator</param>
-    /// <param name="key">The device key to verify</param>
-    /// <returns>True if the key is correct</returns>
-    bool IsDeviceKeyCorrect(DeviceLocator locator, DeviceKey key);
-
-    /// <summary>
-    /// Checks if a device is authorized
-    /// </summary>
-    /// <param name="device">The device locator</param>
-    /// <returns>True if the device is authorized</returns>
-    bool IsDeviceAuthorized(DeviceLocator device);
-
-    /// <summary>
-    /// Gets the private device key for local device
-    /// </summary>
-    /// <returns>The private device key, or null if not available</returns>
-    DeviceKey? GetPrivateDeviceKey();
-
     /// <summary>
     /// Encrypts a string
     /// </summary>
@@ -84,7 +30,7 @@ public interface ISecurityService
     /// <param name="key">The device key containing the public key</param>
     /// <param name="data">The data to encrypt</param>
     /// <returns>The encrypted data as Base64 string</returns>
-    string? RsaEncryptString(DeviceKey key, string data);
+    string? RsaEncryptString(Shared.CSharp.Device.DeviceKey key, string data);
 
     /// <summary>
     /// Decrypts a string using RSA with a specific device's private key
@@ -92,7 +38,7 @@ public interface ISecurityService
     /// <param name="key">The device key containing the private key</param>
     /// <param name="encryptedData">The encrypted data as Base64 string</param>
     /// <returns>The decrypted data</returns>
-    string? RsaDecryptString(DeviceKey key, string encryptedData);
+    string? RsaDecryptString(Shared.CSharp.Device.DeviceKey key, string encryptedData);
 
     /// <summary>
     /// Encrypts content using RSA+AES hybrid encryption
@@ -100,7 +46,7 @@ public interface ISecurityService
     /// <param name="key">The device key</param>
     /// <param name="content">The content to encrypt</param>
     /// <returns>The encrypted content</returns>
-    EncryptedContent RsaEncryptContent(DeviceKey key, string content);
+    EncryptedContent RsaEncryptContent(Shared.CSharp.Device.DeviceKey key, string content);
 
     /// <summary>
     /// Decrypts content using RSA+AES hybrid decryption
@@ -108,7 +54,7 @@ public interface ISecurityService
     /// <param name="key">The device key</param>
     /// <param name="content">The encrypted content</param>
     /// <returns>The decrypted content</returns>
-    string RsaDecryptContent(DeviceKey key, EncryptedContent content);
+    string RsaDecryptContent(Shared.CSharp.Device.DeviceKey key, EncryptedContent content);
 
     /// <summary>
     /// Encrypts a string with AES
@@ -126,13 +72,6 @@ public interface ISecurityService
     /// <param name="isSourceInBase64">Whether the source is in Base64</param>
     /// <returns>The decrypted string</returns>
     string AesDecrypt(string source, string key, bool isSourceInBase64 = true);
-
-    /// <summary>
-    /// Computes a hash
-    /// </summary>
-    /// <param name="content">The content to hash</param>
-    /// <returns>The hash</returns>
-    string ComputeHash(string content);
 
     /// <summary>
     /// Computes SHA1 hash of a string
