+++ KitX Core Contracts/KitX.Core.Contract/Workflow/Nodes/ConstNode.cs
@@ -1,0 +1,38 @@
+namespace KitX.Core.Contract.Workflow;
+
+/// <summary>
+/// Const node - constant value
+/// </summary>
+public class ConstNode : BlueprintNode
+{
+    /// <summary>
+    /// Constant name
+    /// </summary>
+    public string ConstName { get; set; } = string.Empty;
+
+    /// <summary>
+    /// Constant type
+    /// </summary>
+    public string ConstType { get; set; } = "string";
+
+    /// <summary>
+    /// Constant value
+    /// </summary>
+    public string? ConstValue { get; set; }
+
+    public ConstNode()
+    {
+        NodeType = BlueprintNodeType.Const;
+        Name = "Const";
+        InitializePinsFromDescriptor();
+    }
+
+    public override NodeDescriptor GetDescriptor() => new(
+        Width: 120, Height: 50,
+        InputPins: [],
+        OutputPins: [new PinDescriptor("Value", PinType.Any, 25)],
+        DisplayName: "Const"
+    );
+
+    public override string GetDisplayTitle() => $"Const: {ConstName}";
+}
