+++ KitX Core Contracts/KitX.Core.Contract/Workflow/Nodes/PinDescriptor.cs
@@ -1,0 +1,11 @@
+namespace KitX.Core.Contract.Workflow;
+
+/// <summary>
+/// Describes a pin's layout within a node template.
+/// Used for self-describing node pin configurations.
+/// </summary>
+public record PinDescriptor(
+    string Name,
+    PinType Type,
+    double RelativeY
+);
