+++ KitX Core Contracts/KitX.Core.Contract/Workflow/Nodes/PinDirection.cs
@@ -1,0 +1,10 @@
+namespace KitX.Core.Contract.Workflow;
+
+/// <summary>
+/// Pin direction
+/// </summary>
+public enum PinDirection
+{
+    Input,
+    Output
+}
