+++ KitX Core Contracts/KitX.Core.Contract/Workflow/Statements/ExpressionStatement.cs
@@ -1,0 +1,12 @@
+namespace KitX.Core.Contract.Workflow;
+
+/// <summary>
+/// Expression statement
+/// </summary>
+public class ExpressionStatement : BlockStatement
+{
+    /// <summary>
+    /// The expression to execute
+    /// </summary>
+    public string Expression { get; set; } = string.Empty;
+}
