+++ KitX Core Contracts/KitX.Core.Contract/Workflow/Statements/VariableDeclarationStatement.cs
@@ -1,0 +1,12 @@
+namespace KitX.Core.Contract.Workflow;
+
+/// <summary>
+/// Variable declaration statement
+/// </summary>
+public class VariableDeclarationStatement : BlockStatement
+{
+    /// <summary>
+    /// The variable declaration
+    /// </summary>
+    public VariableDeclaration Declaration { get; set; } = new();
+}
